refactor: hide header on mobile screens, remove set command from scripts and move performance files to router
