+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "clean": "rimraf dist",
-    "dev": "set NODE_ENV=development&& vite",
-    "build": "set NODE_ENV=production&& npm run clean && vite build",
+    "dev": "vite",
+    "build": "npm run clean && vite build",
     "preview": "vite preview",
     "test": "vitest",
     "test:ci": "vitest run --coverage --silent --ci",
