+++ src/pages/Home/index.ts
@@ -1,4 +1,4 @@
-import LazyLoader from "@/performance";
+import LazyLoader from "@/router/LazyLoader";
 import HomeLoading from "./Home.loading";
 
 const HomePage = LazyLoader(() => import("./Home"), HomeLoading);
