+++ src/router/index.ts
@@ -1,3 +1,5 @@
 import Routes from "./AppRoutes";
+import LazyLoader from "./LazyLoader";
 
 export default Routes;
+export { LazyLoader };
